+++ two-pointers/532.k-diff-pairs-in-an-array-s2.cpp
@@ -6,8 +6,8 @@ public:
   int findPairs(std::vector<int> &nums, int k) {
     std::sort(nums.begin(), nums.end());
 
-    int i = 0;
-    int j = 0;
+    size_t i = 0;
+    size_t j = 0;
     int count = 0;
 
     while (j < nums.size()) {
