+++ two-pointers/713.subarray-product-less-than-k.cpp
@@ -4,10 +4,10 @@ class Solution {
 public:
   int numSubarrayProductLessThanK(std::vector<int> &nums, int k) {
     if (k <= 1) return 0;
-    int j = 0;
+    size_t j = 0;
     int prod = 1;
     int count = 0;
-    for (int i = 0; i < nums.size(); i++) {
+    for (size_t i = 0; i < nums.size(); i++) {
       if (j < i) {
         j = i;
         prod = 1;
