+++ two-pointers/80.remove-duplicates-from-sorted-array-ii.cpp
@@ -7,7 +7,7 @@ public:
     if (nums.size() == 1) return 1;
 
     int i = 1;
-    for (int j = 0; j < nums.size(); j++) {
+    for (size_t j = 0; j < nums.size(); j++) {
       if (j == 0 || j == 1)
         continue;
       else if (nums[j] == nums[i] && nums[j] == nums[i - 1])
